Add files via upload
So imagine a cohort of people with an infectious disease. It's a simple disease like simple flu disease.  So we're going to assume that ultimately, everyone recovers from the infection after a certain infectious period. We can then separate the  cohort into two compartments. One is the number of people still infected with the disease and the other is the number of people who have recovered. We start with everyone in the infected compartment and end up with everyone in the recovered compartment. So far, it is simple. But how long does it take for people to move from one compartment to another? If we knew the infectious period, then could we predict how many people would be infected and recovered after one week, after a month, or how about after a year? To model this, we will use the rate of transition between I (infectious) and R (recovered). Many of you may be familiar with constant hazard rates and survival analysis. What we are doing here is basically the same thing. So first, you can model the dynamics of this cohort using just two differential equations: dI by dt equals minus Gamma times I and dR by dt equals Gamma times I. You'll recognize these as simple differential equations, where dI by dt denotes the rate of change of I with respect to time. We've just seen that the rate of flow out of the I compartment is proportional to the number of people in the I compartment, and the constant of proportionality is Gamma. We call Gamma the recovery rate. So the larger the value of Gamma, the more quickly people go from I to R. That is, the more quickly they recover. We'll make this a bit more precise very soon, and we'll describe how Gamma actually relates to the infectious period. But it turns out that this model is simple enough that we can actually write down the solution to it. Now, this is a very simple system of equations, but it's important to realize that we've already made a series of assumptions to get here. Perhaps most important is that at any given point in time, every individual in the I compartment is just as likely to experience a recovery. This is why we're able to attach this rate Gamma to everyone in the I compartment.  So next, let's look at the behavior of I and R over time for different values of Gamma. We will start with a cohort of 1,000 people and simulating how they recover over time. Note that Gamma is a rate per unit time. So it has units of per day or day to the minus one. This graph has time in days along the x-axis. You can see that when we choose a value of Gamma that is 0.1 per day, some people in the cohort recover quickly, others recover slowly, but on average, it takes people 10 days to recover.  When we have a higher value of Gamma, say 0.5 per day, on average, it takes people two days to recover. So what's going on here is the behavior of the exponential distribution. It turns out that for any population governed by this assumption, the time spent in the I compartment is distributed according to an exponential distribution with exponential parameter Gamma. The mean of that distribution or the mean infectious period is just one over Gamma. It makes sense, right? The shorter the infectious period, the shorter the duration, and therefore, the larger the value of Gamma. So to recap, we have just written down a simple model to describe the dynamics of an infected cohort. As long as you know the value of Gamma, you'll be able to say how many people are still infected and how many people are recovered at any given point of time. Here we've talked about the infectious period and recovery times. But now imagine more complex models that involve additional compartments and rates to describe other types of transitions like mortality. In all of these models, two important things to remember about rates like Gamma are that they are in units of inverse time, so they could be days to the minus one or even years to the minus one, and secondly, the inverse of the rate is the average duration that people spend in a given compartment.